--- UserWebGUI/html/CrvDqm.html
@@ -1,127 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>CRV DQM</title>
-  <script src="/jsrootsys/scripts/JSRoot.core.js"></script>
-  <style>
-    body  { font-family: sans-serif; background: #f5f5f5; color: #333; margin: 16px; }
-    h1    { font-size: 1.2em; margin-bottom: 12px; }
-    #status { font-size: 0.85em; color: #666; margin-bottom: 8px; }
-    #controls { margin-bottom: 8px; }
-    #controls button {
-      font-size: 0.85em; padding: 4px 10px; margin-right: 6px; cursor: pointer;
-      background: #fff; border: 1px solid #bbb; border-radius: 3px;
-    }
-    #controls button:hover { background: #e8e8e8; }
-    #canvas_wrap { width: 100%; max-width: 1400px; }
-    #canvas_div { width: 100%; height: 900px; background: #fff; }
-    #canvas_wrap:fullscreen { max-width: none; width: 100vw; height: 100vh; background: #fff; padding: 8px; box-sizing: border-box; }
-    #canvas_wrap:fullscreen #canvas_div { height: calc(100vh - 20px); }
-  </style>
-</head>
-<body>
-  <h1>CRV DQM</h1>
-  <div id="status">Loading...</div>
-  <div id="controls">
-    <button id="btn_fullscreen" type="button">Fullscreen</button>
-    <button id="btn_snapshot" type="button">Download PNG</button>
-  </div>
-  <div id="canvas_wrap">
-    <div id="canvas_div"></div>
-  </div>
-  <script>
-    let updating = false;
-    let lastPainter = null;
-
-    function toggleFullscreen() {
-      let wrap = document.getElementById('canvas_wrap');
-      if (!document.fullscreenElement) {
-        wrap.requestFullscreen().catch(err => {
-          document.getElementById('status').textContent =
-            'Fullscreen failed: ' + err.message;
-        });
-      } else {
-        document.exitFullscreen();
-      }
-    }
-
-    async function downloadSnapshot() {
-      try {
-        if (!lastPainter) return;
-        let svg = await lastPainter.produceImage(true, 'svg');
-        if (!svg) throw new Error('no image produced');
-        let img = new Image();
-        let blob = new Blob([svg], { type: 'image/svg+xml;charset=utf-8' });
-        let url = URL.createObjectURL(blob);
-        img.onload = function() {
-          let cnv = document.createElement('canvas');
-          cnv.width = img.width || 1600;
-          cnv.height = img.height || 1000;
-          let ctx = cnv.getContext('2d');
-          ctx.fillStyle = '#fff';
-          ctx.fillRect(0, 0, cnv.width, cnv.height);
-          ctx.drawImage(img, 0, 0);
-          URL.revokeObjectURL(url);
-          let a = document.createElement('a');
-          a.download = 'CrvDqm_' + new Date().toISOString().replace(/[:.]/g, '-') + '.png';
-          a.href = cnv.toDataURL('image/png');
-          a.click();
-        };
-        img.onerror = function() {
-          URL.revokeObjectURL(url);
-          document.getElementById('status').textContent = 'Snapshot failed: image load error';
-        };
-        img.src = url;
-      } catch(e) {
-        document.getElementById('status').textContent = 'Snapshot failed: ' + e.message;
-      }
-    }
-
-    function setPalette(id) {
-      if (JSROOT.settings) JSROOT.settings.Palette = id;
-      if (globalThis.JSROOT && globalThis.JSROOT.settings) globalThis.JSROOT.settings.Palette = id;
-    }
-
-    async function update() {
-      if (updating) return;
-      updating = true;
-      try {
-        // kInvertedDarkBodyRadiator = 56 in ROOT palette IDs
-        setPalette(56);
-        let canvas = await JSROOT.httpRequest('/CrvOtsDqmDisplay/root.json', 'object');
-        document.getElementById('canvas_div').innerHTML = '';
-        lastPainter = await JSROOT.draw('canvas_div', canvas);
-        document.getElementById('status').textContent =
-          'Last update: ' + new Date().toLocaleTimeString();
-      } catch(e) {
-        document.getElementById('status').textContent =
-          'Update failed: ' + e.message + ' \u2014 retrying...';
-      } finally {
-        updating = false;
-      }
-    }
-
-    async function init() {
-
-      document.getElementById('btn_fullscreen').addEventListener('click', toggleFullscreen);
-      document.getElementById('btn_snapshot').addEventListener('click', downloadSnapshot);
-
-      let refreshMs = 500;
-      try {
-        let txt = await JSROOT.httpRequest('/config/refreshMs/cmd.json?method=GetTitle', 'text');
-        let val = parseFloat(JSON.parse(txt).value);
-        if (val > 0) refreshMs = val;
-      } catch(e) {}
-
-      // setTimeout chain prevents overlapping redraws
-      async function loop() {
-        await update();
-        setTimeout(loop, refreshMs);
-      }
-      loop();
-    }
-    init();
-  </script>
-</body>
-</html>
