+++ otsdaq-mu2e-crv/ArtModules/CMakeLists.txt
@@ -11,7 +11,7 @@ cet_build_plugin(CrvVstDemoViewer art::module LIBRARIES
     artdaq::DAQdata
 )
 
-cet_build_plugin(CrvOtsDqm art::module LIBRARIES REG
+cet_build_plugin(CrvDQM art::module LIBRARIES REG
     art_root_io::TFileService_service
     artdaq-core-mu2e::Data
     otsdaq-mu2e::ArtModules
