--- otsdaq-mu2e-crv/fcl/RunCrvOtsDqm.fcl
@@ -1,60 +1,0 @@
-#include "mu2e-trig-config/core/trigProducers.fcl"
-#include "Offline/fcl/standardServices.fcl"
-#include "Offline/CRVResponse/fcl/prolog.fcl"
-
-process_name : CrvOtsDqm
-
-services :{
-  @table::Services.Core
-  @table::Services.Reco
-  TFileService : {
-    fileName : "crvOtsDqm.root"
-  }
-}
-
-source : {
-  module_type : RootInput
-  fileNames   : @nil
-  maxEvents   : -1
-}
-
-physics : {
-  producers :
-  {
-    crvdigi :
-    {
-      module_type   : CrvDigisFromArtdaqFragmentsFEBII
-      diagLevel     : 0
-      useSubsystem0 : false
-      produceNZS : false
-      # useOnlineChannel: true
-
-    }
-  }
-  analyzers : {
-    crvdqm : {
-      module_type           : CrvOtsDqm
-      diagLevel             : 1
-      sendHists             : false
-      dummyHist             : false
-    }
-    CrvDQMcollector: @local::CrvDQMcollector
-  }
-  t1 : [ crvdigi ]
-  a1 : [ crvdqm ]
-  trigger_paths  : [t1]
-  end_paths      : [a1]
-}
-
-outputs: {
-  # out1: {
-  #   module_type: FileDumperOutput
-  #   wantProductFriendlyClassName: true   # Print class names
-  #   onlyIfPresent: true # Only present products
-  # }
-}
-
-services.TimeTracker.printSummary: false
-services.scheduler.wantSummary: false
-services.message.destinations.log.outputStatistics : false
-services.GeometryService.inputFile: "Offline/Mu2eG4/geom/geom_common_extracted.txt"
